Move schema service tests
